add user with roles, validation and add jwt implementations

+++ src/routes/project.ts
@@ -1,10 +1,14 @@
+import { Role } from "@prisma/client";
 import { Router } from "express";
 import {
   createNewProject,
   deleteProjectsById,
+  getProjectsById,
   getProjectsByPage,
   updateProjectById,
 } from "../services/project";
+import { getUserRoleById } from "../services/user";
+import { UnauthorizedError } from "../types/error";
 import { validateDelete } from "../validators/common";
 import {
   validateCreateProject,
@@ -16,7 +20,8 @@ const projectRouter = Router();
 projectRouter.get("/:page", async (req, res, next) => {
   try {
     const page = +req.params.page ?? 0;
-    res.send(await getProjectsByPage(page));
+    const user = await getUserRoleById(req.user.id);
+    res.send(await getProjectsByPage(page, user.role));
   } catch (error) {
     next(error);
   }
@@ -25,6 +30,12 @@ projectRouter.get("/:page", async (req, res, next) => {
 projectRouter.post("/", validateCreateProject, async (req, res, next) => {
   try {
     const { project, tasks } = req.body;
+    const user = await getUserRoleById(req.user.id);
+    if (user.role !== Role.admin || project.owner_id !== req.user.id) {
+      throw new UnauthorizedError(
+        "Only admin can create a project on his own id"
+      );
+    }
     res.send(await createNewProject(project, tasks));
   } catch (error) {
     next(error);
@@ -34,6 +45,16 @@ projectRouter.post("/", validateCreateProject, async (req, res, next) => {
 projectRouter.delete("/", validateDelete, async (req, res, next) => {
   try {
     const { ids } = req.body;
+    const [user, projects] = await Promise.all([
+      getUserRoleById(req.user.id),
+      getProjectsById(ids),
+    ]);
+    if (
+      user.role !== Role.admin ||
+      projects.some((project) => project.owner_id !== req.user.id)
+    ) {
+      throw new UnauthorizedError("Only admin can delete his own project");
+    }
     res.send(await deleteProjectsById(ids));
   } catch (error) {
     next(error);
@@ -43,6 +64,11 @@ projectRouter.delete("/", validateDelete, async (req, res, next) => {
 projectRouter.patch("/:id", validateUpdateProject, async (req, res, next) => {
   try {
     const { id } = req.params;
+    const { data } = req.body;
+    const user = await getUserRoleById(req.user.id);
+    if (user.role !== Role.admin || data.owner_id !== req.user.id) {
+      throw new UnauthorizedError("Only admin can update a project");
+    }
     res.send(await updateProjectById(+id, req.body));
   } catch (error) {
     next(error);
